requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.2.0
neo4j>=6.0.0
python-multipart>=0.0.6
